btn components moved to btn folder in components folder
