+++ app/routes/quotes/$quoteId.tsx
@@ -2,13 +2,10 @@ import { Outlet, useActionData, useLoaderData } from "@remix-run/react";
 import { redirect } from "@remix-run/server-runtime";
 import { useState } from "react";
 import AddNoteCard from "~/components/AddNoteCard";
-import EditQuoteBtn from "~/components/EditQuoteBtn";
+import EditQuoteBtn from "~/components/Buttons/EditQuoteBtn";
 import EditQuoteCard from "~/components/EditQuoteCard";
 import PageTitle from "~/components/PageTitle";
-import QuoteBackBtn from "~/components/QuoteBackBtn";
-import QuoteCardLarge from "~/components/QuoteCardLarge";
-import QuoteDetailCard from "~/components/QuoteDetailCard";
-import QuoteCard from "~/components/QuoteDetailCard";
+import QuoteBackBtn from "~/components/Buttons/QuoteBackBtn";
 import QuoteNoteGrid from "~/components/QuoteNoteGrid";
 import QuoteTags from "~/components/QuoteTags";
 import SectionTitle from "~/components/SectionTitle";
@@ -44,143 +41,9 @@ export const loader = async ({params, request}: any) => {
     return {author, quote, content, notes}
 }
 
-export const action = async ({ request, params }: any) => {
-    const userId = await requireUserId(request);
-    const form = await request.formData()
-    const formBody = form.get('body')
-    const quoteBody = form.get('quoteBody')
-    const authorId = form.get('authorId')
-    const contentId = form.get('contentId')
-    const id = params.quoteId
-    const isFavorited = form.get('isFavorited')
-    const tagBody = form.get('tagBody')
-    const date: any = new Date
-    const updatedAt = date.toISOString()
-    console.log(Object.fromEntries(form))
-
-    // Action to delete quote
-    if(form.get('_method') === 'delete') {
-        await prisma.quote.delete({ where: { id: params.quoteId}})
-        return redirect('/quotes')
-    }
-
-    // Action to update quote
-    if(form.get('_method') === 'update') {
-        const body = quoteBody
-
-        const errors = {
-            body: ''
-        }
-    
-        // validation check to make sure body isn't less than 4 characters
-        function checkBody(body: any) {
-            if(!body || body.length < 4) {
-                return errors.body = `Quote too short`
-            }
-        }
-    
-        checkBody(body)
-    
-        if (errors.body) {
-            const values = Object.fromEntries(form)
-            return { errors, values }
-        }
-
-        const fields = {body}
-        await prisma.quote.update({where: {id: params.quoteId}, data: fields})
-        return redirect('/quotes')
-    }
-
-
-    // Action to create a quoteNote
-    if(form.get('_method') === 'note') {
-        const body = formBody
-        const quoteId = params.quoteId
-
-        const errors = {
-            noteBody: ''
-        }
-    
-        function checkBody(body: any) {
-            if(!body || body.length < 4) {
-                return errors.noteBody = `Note too short`
-            }
-        }
-    
-        checkBody(body)
-    
-        if (errors.noteBody) {
-            const values = Object.fromEntries(form)
-            return { errors, values }
-        }
-
-        const fields = { body, quoteId, userId, authorId, contentId}
-        await prisma.quoteNote.create({ data: fields })
-
-        // update the createdAt date when a new note is added to a quote
-        await prisma.quote.update({
-            where: {id: quoteId},
-            data: { updatedAt: updatedAt}
-        })
-
-        return redirect(`/quotes/${params.quoteId}`)
-    }
-
-    // Action to add tag
-    if(form.get('_method') === 'tag') {
-        const oldBody: string = tagBody
-        const quoteId: string = params.quoteId
-
-        const errors = {
-            tagBody: ''
-        }
-
-        const body = oldBody.replace(/\s+/g, '-').toLowerCase()
-
-        function checkBody(body: any) {
-            if(!body || body.length < 3) {
-                return errors.tagBody = `Tag too short`
-            }
-        }
-    
-        checkBody(body)
-
-    
-        if (errors.tagBody) {
-            const values = Object.fromEntries(form)
-            return { errors, values }
-        }
-
-        const fields = {body, quoteId, userId}
-        await prisma.tag.create({ 
-            data: fields
-        })
-        return redirect(`/quotes/${id}`)
-    }
-    
-    // Action to delete tag
-    if(form.get('_method') === 'deleteTag') {
-        const values = Object.fromEntries(form)
-        const tagId = form.get('tagId')  
-        await prisma.tag.delete({ where: { id: tagId}})
-        return redirect(`/quotes/${id}`)
-    }
-
-    // Action to update favorite status of quote
-    if(form.get('_method') !== ('delete' || 'update' || 'note' || 'deleteTag' || 'tag') ) {
-        await prisma.quote.update({
-            where: { id: id },
-            data: { isFavorited: isFavorited }
-        })
-        return redirect(`/quotes/${id}`)
-    }
-
-    
-}
 
 export default function QuoteDetail() {
     const quote = useLoaderData()
-    const actionData = useActionData()
     const [edit, setEdit] = useState(false)
     console.log('quoteId route --> ', quote)
     return (
