+++ app/routes/quotes/index.tsx
@@ -6,8 +6,9 @@ import QuoteIndexCard from "~/components/QuoteIndexCard";
 import { prisma } from "~/db.server";
 import { requireUserId } from "~/session.server";
 
-import AddQuoteBtn from "~/components/AddQuoteBtn";
+
 import { useState } from "react";
+import AddQuoteBtn from "~/components/Buttons/AddQuoteBtn";
 
 export const loader = async ({request}: any) => {
     const userId = await requireUserId(request);
